Fix (???) frame latency when vsync is enabled
I'm not actually sure how it got fixed. There is apparently nothing actually making us wait for the last frame to be presented before returning from AcquireImage (we're always waiting on a new fence), so the only thing that *should* be causing us to wait is the vkAcquireImage call itself.

We're no longer enabling PRESENT_MODE_FIFO_LATEST_READY even when its available, because we (apparently?) don't throttle execution with it, at least in NeoDemo, negating its primary benefit. (Or else NVIDIA's implementation is somehow wrong.)